moved methods for setting event data to EventBuilder class
