+++ modules/http-statistics-monitoring/src/main/java/org/wso2/appserver/monitoring/config/ConfigurationException.java
@@ -1,8 +1,5 @@
 package org.wso2.appserver.monitoring.config;
 
-/**
- * Created by nathasha on 1/11/16.
- */
 public class ConfigurationException extends Exception {
 
     public ConfigurationException() {
